Substrait Compiler

Substrait Compiler

Substrait Compiler

Details

The SubstraitCompiler defines a mutable object that accumulates information needed to evaluate a substrait.Rel tree. In addition to the substrait.Rel tree itself, the compiler must keep track of function identifiers, column names, and the R objects (e.g., data frames) that will be used as leaf nodes when the plan is evaluated. Specific consumers will need to subclass the SubstraitCompiler and implement the $evaluate() and/or $resolve_function() methods. Typically users will not interact with R6 methods but will use the pipeable interface (e.g. substrait_select()). The pipeable interface clones the compiler before it is modified to minimize the user's interaction to R6 reference semantics.

Get a function reference identifier for a given function/input argument combination.

Public fields

rel

The root of the current substrait.Rel tree.

schema

A substrait.NamedStruct containing the field names and field types of rel.

mask

A named list of substrait.Expression objects where the names are identical to the field names as provided in schema. This list is used as the data mask when evaluating expressions (e.g., rlang::eval_tidy()).

groups

A named list of substrait.Expression to be used for future grouping (e.g., after calling dplyr::group_by()).

Methods

Public methods

SubstraitCompiler$new()

SubstraitCompiler$add_relation()

SubstraitCompiler$named_table()

SubstraitCompiler$named_table_list()

SubstraitCompiler$validate()

SubstraitCompiler$plan()

SubstraitCompiler$evaluate()

SubstraitCompiler$resolve_function()

SubstraitCompiler$function_id()

SubstraitCompiler$function_extension()

SubstraitCompiler$next_id()

SubstraitCompiler$clone()

Method new()

Create a new SubstraitCompiler.

Usage

SubstraitCompiler$new(object = NULL, ...)

Arguments

object

An object, most commonly a data.frame or table-like object.

...

Passed to add_relation() if object is not NULL

Method add_relation()

Sets the rel of this compiler to a substrait.Rel (usually a substrait.Rel.ReadRel) and sets schema and mask to represent the root of the relation tree.

Usage

SubstraitCompiler$add_relation(object, ...)

Arguments

object

An object, most commonly a data.frame or table-like object.

...

Unused by the default method

Method named_table()

Retrieve a named table

Usage

SubstraitCompiler$named_table(name)

Arguments

name

A table name

Returns

The object that was passed

Method named_table_list()

Retrieve all named tables as a list()

Usage

SubstraitCompiler$named_table_list()

Returns

a named list() of objects

Method validate()

Validates a compiler after it was modified. This is an opportunity to provide meaningful feedback (e.g., errors, warnings)

Usage

SubstraitCompiler$validate()

Returns

self

Method plan()

Assembles a substrait.Plan from the current information available to the compiler.

Usage

SubstraitCompiler$plan()

Returns

A substrait.Plan

Method evaluate()

Evaluates the plan being built by the compiler.

Usage

SubstraitCompiler$evaluate(...)

Arguments

...

Extra arguments specific to the compiler type.

Returns

A table-like object whose structure is defined by the SubstraitCompiler class. The returned object should have a as.data.frame() method.

Method resolve_function()

Resolves an R function call as a Substrait function call.

Usage

SubstraitCompiler$resolve_function(name, args, template, output_type = NULL)

Arguments

name

The fully-qualified name of the function as it was called (e.g., pkg::fun). If no package name was explicitly specified, the package name will not be present in name.

args

A list() of arguments. These may be R objects or Substrait objects created while evaluating the user-provided arguments (e.g., field references or function calls).

template

A substrait.Expression.ScalarFunction, a substrait.Expression.WindowFunction, or a substrait.AggregateFunction.

output_type

An explicit output type to use or a function accepting one type per args.

Returns

A modified template with function_reference, args, and output_type set.

Method function_id()

Usage

SubstraitCompiler$function_id(name, arg_types)

Arguments

name

The fully-qualified name of the function as it was called (e.g., pkg::fun). If no package name was explicitly specified, the package name will not be present in name.

arg_types

A list() of substrait.Type objects.

Returns

An integer function reference

Method function_extension()

Retrieve a function extension by anchor/reference

Usage

SubstraitCompiler$function_extension(id)

Arguments

id

An function_anchor/function_reference identifier

Returns

A substrait.extensions.SimpleExtensionDeclaration.ExtensionFunction.

Method next_id()

Get the next unique identifier.

Usage

SubstraitCompiler$next_id()

Returns

An integer that has not been returned by a previous call to next_id() for this instance.

Method clone()

The objects of this class are cloneable with this method.

Usage

SubstraitCompiler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.