dplyr verb implementations

dplyr verb implementations

Usage

# S3 method for class 'SubstraitCompiler'
select(.data, ...)

# S3 method for class 'SubstraitCompiler'
rename(.data, ...)

# S3 method for class 'SubstraitCompiler'
rename_with(.data, .fn, .cols = dplyr::everything(), ...)

# S3 method for class 'SubstraitCompiler'
filter(.data, ...)

# S3 method for class 'SubstraitCompiler'
mutate(.data, ..., .keep = c("all", "used", "unused", "none"))

# S3 method for class 'SubstraitCompiler'
transmute(.data, ...)

# S3 method for class 'SubstraitCompiler'
arrange(.data, ..., .by_group = FALSE)

# S3 method for class 'SubstraitCompiler'
group_by(.data, ..., .add = FALSE, .drop = NULL)

# S3 method for class 'SubstraitCompiler'
ungroup(x, ...)

# S3 method for class 'SubstraitCompiler'
summarise(.data, ..., .groups = NULL)

# S3 method for class 'SubstraitCompiler'
summarize(.data, ..., .groups = NULL)

# S3 method for class 'SubstraitCompiler'
collect(x, ...)

# S3 method for class 'SubstraitCompiler'
relocate(.data, ..., .before = NULL, .after = NULL)

# S3 method for class 'SubstraitCompiler'
semi_join(x, y, by = NULL, ...)

# S3 method for class 'SubstraitCompiler'
anti_join(x, y, by = NULL, ...)

# S3 method for class 'SubstraitCompiler'
inner_join(x, y, by = NULL, suffix = c(".x", ".y"), ..., keep = NULL)

# S3 method for class 'SubstraitCompiler'
left_join(x, y, by = NULL, suffix = c(".x", ".y"), ..., keep = NULL)

# S3 method for class 'SubstraitCompiler'
right_join(x, y, by = NULL, suffix = c(".x", ".y"), ..., keep = NULL)

# S3 method for class 'SubstraitCompiler'
full_join(x, y, by = NULL, suffix = c(".x", ".y"), ..., keep = NULL)

# S3 method for class 'SubstraitCompiler'
distinct(.data, ..., .keep_all = FALSE)

# S3 method for class 'SubstraitCompiler'
count(.data, ..., wt = NULL, sort = FALSE, name = NULL)

Arguments

.data, x, y

A substrait_compiler()

...

select(): see dplyr::select()

rename(): see dplyr::rename()

filter(): see dplyr::filter()

mutate(): see dplyr::mutate()

arrange(): see dplyr::arrange()

count(): see dplyr::count()

distinct(): see dplyr::distinct()

.fn

Function to transform selected .cols; seedplyr::rename_with()

.cols

Columns to rename; seedplyr::rename_with()

.keep

Which columns are retained in output; see dplyr::mutate()

.by_group

sort by grouping variable? seedplyr::arrange()

.add

Use TRUE to add the groupings to the current groupings and FALSE to reset the grouping.

.drop

Not supported, see dplyr::group_by()

.groups

One of "drop_last", "drop", or "keep" (see dplyr::summarise()).

.before

Destination of columns to move; see dplyr::relocate()

.after

Destination of columns to move; see dplyr::relocate()

by

For joins, a join specifier or NULL to use common variables across x and y; see dplyr::inner_join().

suffix

A suffix used to disambiguate columns from x and y if a join would result in duplicate column names.

keep

For joins, use TRUE to keep all output columns; see dplyr::inner_join().

.keep_all

If TRUE, keep all variables in .data; see dplyr::distinct()

wt

Frequency weights; see dplyr::count()

sort

If TRUE, will show the largest groups at the top; see dplyr::count()

name

The name of the new column in the output; see dplyr::count()

Value

A modified substrait_compiler()

Examples

if (FALSE) { # has_duckdb_with_substrait()
library(dplyr)
compiler <- duckdb_substrait_compiler(mtcars)

select(compiler, mpg2 = mpg) %>% collect()
rename(compiler, mpg2 = mpg) %>% collect()
filter(compiler, mpg > 20) %>% collect()
mutate(compiler, mpg + 10) %>% collect()
transmute(compiler, mpg + 10) %>% collect()
arrange(compiler, desc(mpg)) %>% collect()
}